feat: support astro 6
