+++ .changeset/serious-mayflies-greet.md
@@ -3,3 +3,5 @@
 ---
 
 Adds support for Astro 6.0
+
+Support for Astro 5.0 is dropped, use v3 of astro-sst by pinning to `astro-sst@three`.
